moving more logic related to privacy shield anim into module
