+++ app/src/main/java/com/duckduckgo/app/global/model/Site.kt
@@ -25,7 +25,7 @@ import com.duckduckgo.app.global.model.SiteFactory.SitePrivacyData
 import com.duckduckgo.app.privacy.model.HttpsStatus
 import com.duckduckgo.app.privacy.model.PrivacyGrade
 import com.duckduckgo.app.privacy.model.PrivacyPractices
-import com.duckduckgo.app.privacy.model.PrivacyShield
+import com.duckduckgo.privacy.dashboard.api.PrivacyShield
 import com.duckduckgo.app.surrogates.SurrogateResponse
 import com.duckduckgo.trackerdetection.model.Entity
 import com.duckduckgo.app.trackerdetection.model.TrackingEvent
