+++ privacy-dashboard/privacy-dashboard-api/src/main/java/com/duckduckgo/privacy/dashboard/api/animations/BrowserTrackersAnimatorHelper.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.privacy.dashboard.api
+package com.duckduckgo.privacy.dashboard.api.animations
 
 import android.view.View
 import android.view.ViewGroup
