+++ privacy-dashboard/privacy-dashboard-impl/src/main/res/raw/dark_protected_shield.json
