+++ privacy-dashboard/privacy-dashboard-impl/src/main/res/raw/dark_unprotected_shield.json
