+++ privacy-dashboard/privacy-dashboard-impl/src/main/res/raw/protected_shield.json
