+++ privacy-dashboard/privacy-dashboard-impl/src/main/res/raw/unprotected_shield.json
